requirements.txt
streamlit>=1.36
python-docx>=1.1
Pillow>=10.4
opencv-python-headless>=4.10
numpy>=1.26
lxml>=5.2
google-genai>=0.7
./template_build
./moore_build
